+++ src/Castle.Core/DynamicProxy/Generators/Emitters/MethodEmitter.cs
@@ -137,8 +137,8 @@ namespace Castle.DynamicProxy.Generators.Emitters
 		{
 			if (ImplementedByRuntime == false && CodeBuilder.IsEmpty)
 			{
-				CodeBuilder.AddStatement(new NopStatement());
-				CodeBuilder.AddStatement(new ReturnStatement());
+				CodeBuilder.Add(new NopStatement());
+				CodeBuilder.Add(ReturnStatement.Instance);
 			}
 		}