+++ src/Castle.Core/DynamicProxy/Generators/Emitters/SimpleAST/LiteralStringExpression.cs
@@ -1,34 +1,36 @@
-// Copyright 2004-2011 Castle Project - http://www.castleproject.org/
-// 
+// Copyright 2004-2018 Castle Project - http://www.castleproject.org/
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.DynamicProxy.Generators.Emitters
+namespace Castle.DynamicProxy.Generators.Emitters.SimpleAST
 {
-	using Castle.DynamicProxy.Generators.Emitters.SimpleAST;
+	using System.Diagnostics;
+	using System.Reflection.Emit;
 
-	public class TypeConstructorEmitter : ConstructorEmitter
+	internal sealed class LiteralStringExpression : Expression
 	{
-		internal TypeConstructorEmitter(AbstractTypeEmitter maintype)
-			: base(maintype, maintype.TypeBuilder.DefineTypeInitializer())
+		private string value;
+
+		public LiteralStringExpression(string value)
 		{
+			Debug.Assert(value != null);
+
+			this.value = value;
 		}
 
-		public override void EnsureValidCodeBlock()
+		public override void Emit(IMemberEmitter member, ILGenerator gen)
 		{
-			if (CodeBuilder.IsEmpty)
-			{
-				CodeBuilder.Add(ReturnStatement.Instance);
-			}
+			gen.Emit(OpCodes.Ldstr, value);
 		}
 	}
 }